apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ralphite-api"
version = "0.1.0"
description = "Ralphite orchestration API"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115,<1",
  "uvicorn>=0.30,<1",
  "sqlalchemy>=2.0,<3",
  "psycopg[binary]>=3.2,<4",
  "pydantic>=2.8,<3",
  "email-validator>=2.2,<3",
  "pydantic-settings>=2.3,<3",
  "pyyaml>=6.0,<7",
  "python-multipart>=0.0.9,<1",
  "passlib[argon2]>=1.7,<2",
  "PyJWT>=2.9,<3",
  "ralphite-schemas"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9",
  "httpx>=0.27,<1"
]

[project.scripts]
ralphite-api = "ralphite_api.main:run"

[tool.uv.sources]
ralphite-schemas = { workspace = true }

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
